+++ src/Api/Defaut/Event.php
@@ -29,7 +29,6 @@ use LibMelanie\Log\M2Log;
 use LibMelanie\Lib\ICS;
 use LibMelanie\Config\Config;
 use LibMelanie\Config\DefaultConfig;
-use stdClass;
 
 /**
  * Classe evenement par defaut,
@@ -166,25 +165,37 @@ class Event extends MceObject {
    * 
    * @var \DateTime
    */
-  protected $dtstart;
+  protected $_dtstart;
+  /**
+   * DateTime basée sur l'ancien champ $start
+   * 
+   * @var \DateTime
+   */
+  protected $_olddtstart;
   /**
    * DateTime basée sur le champ $start au timezone UTC
    *
    * @var \DateTime
    */
-  protected $dtstart_utc;
+  protected $_dtstart_utc;
   /**
    * DateTime basée sur le champ $end
    *
    * @var \DateTime
    */
-  protected $dtend;
+  protected $_dtend;
+  /**
+   * DateTime basée sur l'ancien champ $end
+   * 
+   * @var \DateTime
+   */
+  protected $_olddtend;
   /**
    * DateTime basée sur le champ $end au timezone UTC
    *
    * @var \DateTime
    */
-  protected $dtend_utc;
+  protected $_dtend_utc;
   
   /**
    * Object VCalendar disponible via le VObject
@@ -1570,7 +1581,7 @@ class Event extends MceObject {
    * @ignore
    *
    */
-  function save($saveAttendees = true) {
+  public function save($saveAttendees = true) {
     M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->save()");
     if (!isset($this->objectmelanie))
       throw new Exceptions\ObjectMelanieUndefinedException();
@@ -1583,6 +1594,10 @@ class Event extends MceObject {
     if ($saveAttendees) {
       $this->saveAttendees();
     }
+    // MANTIS 0007426: Avancer la date de fin de récurrence devrait supprimer les occurrences postérieures
+    if ($this->objectmelanie->fieldHasChanged('enddate')) {
+      $this->deleteOldOccurrences();
+    }
     // Supprimer les exceptions
     if (isset($this->deleted_exceptions) && is_array($this->deleted_exceptions) && count($this->deleted_exceptions) > 0) {
       M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->save() delete " . count($this->deleted_exceptions));
@@ -1591,9 +1606,13 @@ class Event extends MceObject {
         $exception->delete();
       }
     }
+    
     $exMod = false;
     // Sauvegarde des exceptions
     if (isset($this->exceptions)) {
+      // MANTIS 0007427: Modifier toutes les occurrences devrait également modifier les occurrences modifiées si possible
+      $this->updateOccurrences();
+
       foreach ($this->exceptions as $exception) {
         $res = $exception->save();
         $exMod = $exMod || !is_null($res);
@@ -1634,6 +1653,84 @@ class Event extends MceObject {
     M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->save() Rien a sauvegarder: return null");
     return null;
   }
+
+  /**
+   * Lors d'un changement de date de fin de récurrence
+   * supprime toutes les occurrences modifiées après la nouvelle date
+   */
+  protected function deleteOldOccurrences() {
+    $newEndDate = \DateTime::createFromFormat('Y-m-d H:i:s', $this->objectmelanie->enddate, new \DateTimeZone('UTC'));
+    $exceptions = $this->getMapExceptions();
+    if (isset($exceptions)) {
+      foreach ($exceptions as $key => $exception) {
+        if (!$exception->deleted) {
+          $recIdDate = \DateTime::createFromFormat('Y-m-d H:i:s', $exception->recurrence_id, new \DateTimeZone($this->timezone));
+          if ($recIdDate > $newEndDate) {
+            unset($exceptions[$key]);
+          }
+        }
+      }
+      $this->setMapExceptions($exceptions);
+    }
+  }
+
+  /**
+   * Actualise toutes les occurrences avec les nouvelles valeurs 
+   * si elles n'avaient pas changées
+   */
+  protected function updateOccurrences() {
+    $fields = ['title', 'location', 'description', 'status', 'class', 'category'];
+
+    // Gestion de la date
+    if ($this->objectmelanie->fieldHasChanged('start') || $this->objectmelanie->fieldHasChanged('end')) {
+      $oldEventDuration = $this->getMapOlddtend()->diff($this->getMapOlddtstart());
+      $oldStarttime = $this->getMapOlddtstart()->format('H:i:s');
+
+      if ($this->objectmelanie->fieldHasChanged('start')) {
+        $startChangeDuration = $this->getMapOlddtstart()->diff($this->getMapDtstart());
+      }
+
+      if ($this->objectmelanie->fieldHasChanged('end')) {
+        $endChangeDuration = $this->getMapOlddtend()->diff($this->getMapDtend());
+      }
+    }
+
+    foreach ($this->exceptions as $exception) {
+      foreach ($fields as $field) {
+        if ($this->objectmelanie->fieldHasChanged($field) 
+            && $this->objectmelanie->getOldData($field) == $exception->getObjectMelanie()->getFieldValueFromData($field)) {
+          $exception->$field = $this->$field;
+        }
+      }
+
+      // Gestion de la date
+      if (isset($oldEventDuration) && isset($oldStarttime)
+          && $exception->dtstart->format('Y-m-d') == $exception->dtrecurrence_id->format('Y-m-d')
+          && $exception->dtstart->format('H:i:s') == $oldStarttime) {
+
+        $tmpDate = clone $this->getMapOlddtstart();
+        $tmpDate->add($exception->dtstart->diff($exception->dtend));
+
+        M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->updateOccurrences() duration ($exception->realuid) compare = " . ($oldEventDuration == $exDuration));
+
+        if ($tmpDate == $this->getMapOlddtend()) {
+          // on est dans le cas ou l'exception n'a pas changé d'horaire, on va pouvoir appliquer les modifications
+          if (isset($startChangeDuration)) {
+            $dtstart = $exception->dtstart;
+            $dtstart->add($startChangeDuration);
+            $exception->dtstart = $dtstart;
+            $exception->recurrence_id = $exception->start;
+          }
+
+          if (isset($endChangeDuration)) {
+            $dtend = $exception->dtend;
+            $dtend->add($endChangeDuration);
+            $exception->dtend = $dtend;
+          }
+        }
+      }
+    }
+  }
   
   /**
    * Mapping de la suppression de l'objet
@@ -1642,7 +1739,7 @@ class Event extends MceObject {
    * @ignore
    *
    */
-  function delete() {
+  public function delete() {
     M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->delete()");
     if (!isset($this->objectmelanie))
       throw new Exceptions\ObjectMelanieUndefinedException();
@@ -1693,7 +1790,7 @@ class Event extends MceObject {
    * @ignore
    *
    */
-  function load() {
+  public function load() {
     M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->load()");
     $ret = $this->objectmelanie->load();
     if (!$ret && $this->notException())
@@ -1729,7 +1826,7 @@ class Event extends MceObject {
    *          Liste des champs pour lesquels on ne sera pas sensible à la casse
    * @return Event[] Array
    */
-  function getList($fields = [], $filter = "", $operators = [], $orderby = "", $asc = true, $limit = null, $offset = null, $case_unsensitive_fields = []) {
+  public function getList($fields = [], $filter = "", $operators = [], $orderby = "", $asc = true, $limit = null, $offset = null, $case_unsensitive_fields = []) {
     M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->getList()");
     $_events = $this->objectmelanie->getList($fields, $filter, $operators, $orderby, $asc, $limit, $offset, $case_unsensitive_fields);
     if (!isset($_events))
@@ -1918,8 +2015,8 @@ class Event extends MceObject {
     M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->setMapStart()");
     if (!isset($this->objectmelanie)) throw new Exceptions\ObjectMelanieUndefinedException();
     $this->objectmelanie->start = $start;
-    $this->dtstart = null;
-    $this->dtstart_utc = null;
+    $this->_dtstart = null;
+    $this->_dtstart_utc = null;
   }
   
   /**
@@ -1930,7 +2027,7 @@ class Event extends MceObject {
   protected function setMapDtstart($dtstart) {
     M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->setMapDtstart()");
     if (!isset($this->objectmelanie)) throw new Exceptions\ObjectMelanieUndefinedException();
-    $this->dtstart = $dtstart;
+    $this->_dtstart = $dtstart;
     $this->objectmelanie->start = $dtstart->format(self::DB_DATE_FORMAT);
     $this->objectmelanie->timezone = $dtstart->getTimezone()->getName();
   }
@@ -1940,16 +2037,16 @@ class Event extends MceObject {
   protected function getMapDtstart() {
     M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->getMapDtstart()");
     if (!isset($this->objectmelanie)) throw new Exceptions\ObjectMelanieUndefinedException();
-    if (!isset($this->dtstart)) {
+    if (!isset($this->_dtstart)) {
       try {
-        $this->dtstart = new \DateTime($this->objectmelanie->start, new \DateTimeZone($this->getMapTimezone()));
+        $this->_dtstart = new \DateTime($this->objectmelanie->start, new \DateTimeZone($this->getMapTimezone()));
       }
       catch (\Exception $ex) {
         M2Log::Log(M2Log::LEVEL_ERROR, $this->get_class . "->getMapDtstart() Erreur pour l'événement '" . $this->objectmelanie->uid . "' : " . $ex->getMessage());
-        $this->dtstart = new \DateTime();
+        $this->_dtstart = new \DateTime();
       }
     }
-    return $this->dtstart;
+    return $this->_dtstart;
   }
   
   /**
@@ -1958,17 +2055,39 @@ class Event extends MceObject {
   protected function getMapDtstart_utc() {
     M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->getMapDtstart_utc()");
     if (!isset($this->objectmelanie)) throw new Exceptions\ObjectMelanieUndefinedException();
-    if (!isset($this->dtstart_utc)) {
+    if (!isset($this->_dtstart_utc)) {
       try {
-        $this->dtstart_utc = new \DateTime($this->objectmelanie->start, new \DateTimeZone($this->getMapTimezone()));
-        $this->dtstart_utc->setTimezone(new \DateTimeZone('UTC'));
+        $this->_dtstart_utc = new \DateTime($this->objectmelanie->start, new \DateTimeZone($this->getMapTimezone()));
+        $this->_dtstart_utc->setTimezone(new \DateTimeZone('UTC'));
       }
       catch (\Exception $ex) {
         M2Log::Log(M2Log::LEVEL_ERROR, $this->get_class . "->getMapDtstart_utc() Erreur pour l'événement '" . $this->objectmelanie->uid . "' : " . $ex->getMessage());
-        $this->dtstart_utc = new \DateTime();
+        $this->_dtstart_utc = new \DateTime();
       }
     }
-    return $this->dtstart_utc;
+    return $this->_dtstart_utc;
+  }
+
+  /**
+   * Mapping olddtstart field
+   */
+  protected function getMapOlddtstart() {
+    M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->getMapOlddtstart()");
+    if (!isset($this->objectmelanie)) throw new Exceptions\ObjectMelanieUndefinedException();
+    if (!isset($this->_olddtstart)) {
+      try {
+        $start = $this->objectmelanie->getOldData('start');
+        if (!isset($start)) {
+          $start = $this->objectmelanie->start;
+        }
+        $this->_olddtstart = new \DateTime($start, new \DateTimeZone($this->getMapTimezone()));
+      }
+      catch (\Exception $ex) {
+        M2Log::Log(M2Log::LEVEL_ERROR, $this->get_class . "->getMapOlddtstart() Erreur pour l'événement '" . $this->objectmelanie->uid . "' : " . $ex->getMessage());
+        $this->_olddtstart = new \DateTime();
+      }
+    }
+    return $this->_olddtstart;
   }
   
   /**
@@ -1980,8 +2099,8 @@ class Event extends MceObject {
     M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->setMapEnd()");
     if (!isset($this->objectmelanie)) throw new Exceptions\ObjectMelanieUndefinedException();
     $this->objectmelanie->end = $end;
-    $this->dtend = null;
-    $this->dtend_utc = null;
+    $this->_dtend = null;
+    $this->_dtend_utc = null;
   }
   
   /**
@@ -1992,7 +2111,7 @@ class Event extends MceObject {
   protected function setMapDtend($dtend) {
     M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->setMapDtend()");
     if (!isset($this->objectmelanie)) throw new Exceptions\ObjectMelanieUndefinedException();
-    $this->dtend = $dtend;
+    $this->_dtend = $dtend;
     $this->objectmelanie->end = $dtend->format(self::DB_DATE_FORMAT);
     // Pas de timezone ici, il est récupéré dans le dtstart
   }
@@ -2002,16 +2121,16 @@ class Event extends MceObject {
   protected function getMapDtend() {
     M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->getMapDtend()");
     if (!isset($this->objectmelanie)) throw new Exceptions\ObjectMelanieUndefinedException();
-    if (!isset($this->dtend)) {
+    if (!isset($this->_dtend)) {
       try {
-        $this->dtend = new \DateTime($this->objectmelanie->end, new \DateTimeZone($this->getMapTimezone()));
+        $this->_dtend = new \DateTime($this->objectmelanie->end, new \DateTimeZone($this->getMapTimezone()));
       }
       catch (\Exception $ex) {
         M2Log::Log(M2Log::LEVEL_ERROR, $this->get_class . "->getMapDtend() Erreur pour l'événement '" . $this->objectmelanie->uid . "' : " . $ex->getMessage());
-        $this->dtend = new \DateTime();
+        $this->_dtend = new \DateTime();
       }
     }
-    return $this->dtend;
+    return $this->_dtend;
   }
   
   /**
@@ -2020,17 +2139,39 @@ class Event extends MceObject {
   protected function getMapDtend_utc() {
     M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->getMapDtend_utc()");
     if (!isset($this->objectmelanie)) throw new Exceptions\ObjectMelanieUndefinedException();
-    if (!isset($this->dtend_utc)) {
+    if (!isset($this->_dtend_utc)) {
       try {
-        $this->dtend_utc = new \DateTime($this->objectmelanie->end, new \DateTimeZone($this->getMapTimezone()));
-        $this->dtend_utc->setTimezone(new \DateTimeZone('UTC'));
+        $this->_dtend_utc = new \DateTime($this->objectmelanie->end, new \DateTimeZone($this->getMapTimezone()));
+        $this->_dtend_utc->setTimezone(new \DateTimeZone('UTC'));
       }
       catch (\Exception $ex) {
         M2Log::Log(M2Log::LEVEL_ERROR, $this->get_class . "->getMapDtend_utc() Erreur pour l'événement '" . $this->objectmelanie->uid . "' : " . $ex->getMessage());
-        $this->dtend_utc = new \DateTime();
+        $this->_dtend_utc = new \DateTime();
       }
     }
-    return $this->dtend_utc;
+    return $this->_dtend_utc;
+  }
+
+  /**
+   * Mapping olddtend field
+   */
+  protected function getMapOlddtend() {
+    M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class . "->getMapOlddtend()");
+    if (!isset($this->objectmelanie)) throw new Exceptions\ObjectMelanieUndefinedException();
+    if (!isset($this->_olddtend)) {
+      try {
+        $end = $this->objectmelanie->getOldData('end');
+        if (!isset($end)) {
+          $end = $this->objectmelanie->end;
+        }
+        $this->_olddtend = new \DateTime($end, new \DateTimeZone($this->getMapTimezone()));
+      }
+      catch (\Exception $ex) {
+        M2Log::Log(M2Log::LEVEL_ERROR, $this->get_class . "->getMapOlddtend() Erreur pour l'événement '" . $this->objectmelanie->uid . "' : " . $ex->getMessage());
+        $this->_olddtend = new \DateTime();
+      }
+    }
+    return $this->_olddtend;
   }
   
   /**
@@ -2477,7 +2618,7 @@ class Event extends MceObject {
     $_attachments = [];
 
     foreach ($attachments as $attachment) {
-      $_attachment = new stdClass();
+      $_attachment = new \stdClass();
       $_attachment->type = $attachment->type;
 
       if ($attachment->type == Attachment::TYPE_URL) {